apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
pydantic==2.11.7
email-validator==2.2.0
python-dotenv==1.1.1
alembic==1.16.5
